+++ controller/admin_controller.php
@@ -2,7 +2,7 @@
 /**
 *
 * @package phpBB Extension - File Upload
-* @copyright (c) 2017 dmzx - http://www.dmzx-web.net
+* @copyright (c) 2017 dmzx - https://www.dmzx-web.net & martin - https://www.martins-phpbb.com
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -187,6 +187,7 @@ class admin_controller
 			WHERE u.user_id = im.user_id
 			ORDER BY ' . $sql_sort_order;
 		$result = $this->db->sql_query_limit($sql, $number, $start);
+
 		while ($row = $this->db->sql_fetchrow($result))
 		{
 			$file_name = $row['fileupload_realname'];
@@ -204,6 +205,7 @@ class admin_controller
 		$this->db->sql_freeresult($result);
 
 		$base_url = $this->u_action;
+
 		//Start pagination
 		$this->pagination->generate_template_pagination($base_url, 'pagination', 'start', $total_fileupload, $number, $start);