+++ ext.php
@@ -1,0 +1,19 @@
+<?php
+/**
+*
+* @package phpBB Extension - File Upload
+* @copyright (c) 2017 dmzx - https://www.dmzx-web.net & martin - https://www.martins-phpbb.com
+* @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
+*
+*/
+
+namespace dmzx\fileupload;
+
+class ext extends \phpbb\extension\base
+{
+	public function is_enableable()
+	{
+		$config = $this->container->get('config');
+		return version_compare($config['version'], '3.2.0', '>=');
+	}
+}