+++ language/en/common.php
@@ -2,7 +2,7 @@
 /**
 *
 * @package phpBB Extension - File Upload
-* @copyright (c) 2017 dmzx - http://www.dmzx-web.net
+* @copyright (c) 2017 dmzx - https://www.dmzx-web.net & martin - https://www.martins-phpbb.com
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -46,6 +46,8 @@ $lang = array_merge($lang, array(
 	'FILEUPLOAD_NEW_FILENAME'					=> 'File name',
 	'FILEUPLOAD_SUCCEEDED'						=> 'Upload Succeeded!',
 	'FILEUPLOAD_DIRECT_LINK'					=> 'Direct link',
+		'DIRECT_COPY_BUTTON'											=> 'copy',
 	'FILEUPLOAD_URL_LINK'						=> 'URL',
-	'FILEUPLOAD_BY'							=> 'File Upload by',
+		'URL_COPY_BUTTON'												=> 'copy',
+	'FILEUPLOAD_BY'							=> 'File hosted by',
 ));