+++ language/es/common.php
@@ -2,7 +2,7 @@
 /**
 *
 * @package phpBB Extension - File Upload
-* @copyright (c) 2017 dmzx - http://www.dmzx-web.net
+* @copyright (c) 2017 dmzx - https://www.dmzx-web.net & martin - https://www.martins-phpbb.com
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -47,5 +47,6 @@ $lang = array_merge($lang, array(
 	'FILEUPLOAD_SUCCEEDED'					=> '¡Subida correcta!',
 	'FILEUPLOAD_DIRECT_LINK'				=> 'Enlace directo',
 	'FILEUPLOAD_URL_LINK'					=> 'URL',
+	'FILEUPLOAD_FILE_LINK'					=> 'ARCHIVO',
 	'FILEUPLOAD_BY'							=> 'Archivo subido por',
 ));