+++ language/es/info_acp_fileupload.php
@@ -2,7 +2,7 @@
 /**
 *
 * @package phpBB Extension - File Upload
-* @copyright (c) 2017 dmzx - http://www.dmzx-web.net
+* @copyright (c) 2017 dmzx - https://www.dmzx-web.net & martin - https://www.martins-phpbb.com
 * @license http://opensource.org/licenses/gpl-2.0.php GNU General Public License v2
 *
 */
@@ -39,5 +39,4 @@ $lang = array_merge($lang, array(
 	// Log
 	'LOG_FILEUPLOAD_ADD'		=> '<strong>Añadido nuevo archivo</strong><br>» %1$s',
 	'LOG_FILEUPLOAD_SETTINGS'	=> '<strong>Configuración cambiada en Subida de Archivos</strong>',
-	'LOG_FILEUPLOAD_DELETED'	=> '<strong>Archivo borrado</strong><br>» %1$s',
 ));