+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 ## Changelog
 
+### 1.0.3 - 2019-02-22
+
+- Dropped support of 3.1.x
+- Added folder.
+
+### 1.0.2 - 2018-06-30
+
+- Added copy button by martin.
+- Twig update.
+- Update URL in files and copyright.
+
 ### 1.0.1 - 2017-03-20
 
 - Update code.